# Faster diff (for some diffs) by skipping pygit2